distribute functions to library
